[themotte/rDrama#451] Move comment_on_publish() to files/helpers/comments.py
