+++ files/routes/admin.py
@@ -13,7 +13,7 @@ from files.classes import *
 from flask import *
 from files.__main__ import app, cache, limiter
 from .front import frontlist
-from files.routes.comments import comment_on_publish
+from files.helpers.comments import comment_on_publish
 from datetime import datetime
 import requests
 from urllib.parse import quote, urlencode
